sampling/batch/src/requirements.txt
dataclasses # Incase we're not use py >=3.7
numpy
absl-py
gin-config
apache-beam[gcp]

# Assumes torch is installed on system, to side-step 
# that torch and extensions cannot be installed with the same pip command
torch-geometric
networkx




